Separate helper responsibilities
