+++ AutomateWashingtonUploads/AutomateWashingtonUploads.csproj
@@ -80,10 +80,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Completion.cs" />
-    <Compile Include="Helper.cs" />
+    <Compile Include="Helpers\DataHelper.cs" />
+    <Compile Include="Helpers\EmailHelper.cs" />
+    <Compile Include="Helpers\ValidationHelper.cs" />
     <Compile Include="Logger.cs" />
-    <Compile Include="LoginInfo.cs" />
-    <Compile Include="PlumbingCourses.cs" />
+    <Compile Include="StaticData\LoginInfo.cs" />
+    <Compile Include="StaticData\PlumbingCourses.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Uploader.cs" />
@@ -104,6 +106,9 @@
       <ProductName>.NET Framework 3.5 SP1</ProductName>
       <Install>false</Install>
     </BootstrapperPackage>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Data\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\Selenium.Firefox.WebDriver.0.24.0\build\Selenium.Firefox.WebDriver.targets" Condition="Exists('..\packages\Selenium.Firefox.WebDriver.0.24.0\build\Selenium.Firefox.WebDriver.targets')" />
